Make home the dashboard
